Update README.md
This project is a responsive website designed to promote healthy eating and sell nutritious food products. It features an attractive user interface and emphasizes the benefits of a balanced diet. Key functionalities include:

Showcasing a variety of health food products
Informative sections on the importance of healthy eating
A user-friendly layout for seamless browsing and shopping experience
Built using HTML, CSS, and JavaScript, this website provides an engaging platform for customers to explore and purchase health-focused food items.